Move legal consent view model to its own file
