+++ novawallet/Modules/LegalConsent/ViewModel/LegalConsentViewModel.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct LegalConsentViewModel {
+    let title: String
+    let subtitle: String
+    let agreement: NSAttributedString
+    let acceptTitle: String
+    let consentAccepted: Bool
+}
